первый коммит

+++ Prakt15/Models/ProductTag.cs
@@ -5,8 +5,8 @@ namespace Prakt15.Models;
 
 public partial class ProductTag
 {
-    public double? ProductId { get; set; }
-    public double? TagId { get; set; }
+    public int? ProductId { get; set; }
+    public int? TagId { get; set; }
 
     public virtual Product? Product { get; set; }
     public virtual Tag? Tag { get; set; }
